cleanup(ui): remove unused imports and code across screens
- Remove unused imports from screens
- Minor formatting cleanup
- Remove dead code

+++ lib/features/activities/presentation/screens/activities_list_screen.dart
@@ -239,7 +239,7 @@ class _ActivitiesListScreenState extends ConsumerState<ActivitiesListScreen> {
                           return ActivityCard(activity: selectedDateActivities[activityIndex]);
                         }
                       },
-                      childCount: (selectedDateCarbDays.length + selectedDateActivities.length) as int,
+                      childCount: (selectedDateCarbDays.length + selectedDateActivities.length),
                     ),
                   ),
                 ),
